* add user space object type   - has reference to destructor   - every object used in usr space inherit this type   - so every object can be safely closed when its owner thread is destroyed

+++ include/sys/tch_mailq.h
@@ -18,7 +18,7 @@ typedef void* tch_mailqId;
 
 
 struct _tch_mailbox_ix_t {
-	tch_mailqId (*create)(size_t sz,uint32_t qlen);
+	tch_mailqId (*create)(uint32_t sz,uint32_t qlen);
 	void* (*alloc)(tch_mailqId qid,uint32_t millisec,tchStatus* result);
 	void* (*calloc)(tch_mailqId qid,uint32_t millisec,tchStatus* result);
 	tchStatus (*put)(tch_mailqId qid,void* mail);
